Added script for tentacles
It still needs work, right now it pushes in a circle but for some reason isn't accurate.
As sources to calculate the force, I tried with 1 and 3 points.
The idea is that when tentacle appears pushes away other objects with rigidbodies. That's why the function findExplosiones is commented out from the Start function, but it will be triggered every 2 seconds on the Update for testing purposes.