ghost AI and Win Condition
MovingPlayers for Testing is a script to test ghosts. This will move to any player fall off from the road.
Game Manager now holds the ammount of players alive, and once it's equal to 1 it will look for the player alive, debug a message to inform it and pause the game.
Now fallOff script(the only way to die for now) will update the number on the GameManager script, so the game will know the current state of players alive.