Implemented Smell Status Effect into PlayerController
Sometimes when 2+ cars are in radius of the player, all the cars in the scene move regardless whether or not they are in range of the radius. Some carAI drive on the edge of the bridge. *NOTE* isSmelly is set to true on Start(). Change to false to stop this part of the script.