Ghosts working on main scene, player respawns after 5 seconds
However, there is an issue with the health bar, if you comment
out playerHealth.updateHealthBar(hitPoints); line 947 from playerController script will make it work properly. Ghosts issue not getting spawned due to a new script not on all prefabs.(arrangeLayers)

Now fallBackChecker destroys Player after 2 seconds.

MovingPlayersForTesting is a script to test ghosts with less than 4 controllers. This will move the player that holds it to fall off from the road.

Game Manager now holds the ammount of players alive, and once it's equal to 1 it will look for the player alive, debug a message to inform it and pause the game.

To use the ghost feature, just call it using GetComponent<fallOff>().fall();

Another minor bug, m_timeToSpawn4 being called as m_timeToSpawn1 making the 4th player waiting time equals to the 1st one