Difficulty Overtime changes added
This is a script added to the game manager, and from there
will adjust the ammount of things being spawned relative
to time.
New functions added to the fish and car spawner to modify the max and min time to spawn. In case is needed, tentacles can be adjusted easily later on.
Kraken will appear after 1 min
Cars and fish will spawn 0.5 sooner after 45 secs.
The ammount of cars maximum will be higher by 1 after 1 minute.